+++ packages/fabric/editor-core/src/transformers/bitbucket/index.ts
@@ -1,23 +1,24 @@
 import { bitbucketSchema } from '@atlaskit/editor-common';
-import {
-  DOMParser,
-  Node as PMNode,
-  Schema,
-} from 'prosemirror-model';
-import {
-  MarkdownSerializer,
-  marks,
-  nodes,
-} from './serializer';
+import { DOMParser, Node as PMNode, Schema } from 'prosemirror-model';
+import { MarkdownSerializer, marks, nodes } from './serializer';
 import { Transformer } from '../transformer';
 import { transformHtml } from './util';
+
+export interface TransformerOptions {
+  disableBitbucketLinkStripping?: boolean;
+}
 
 export default class BitbucketTransformer implements Transformer<string> {
   private serializer = new MarkdownSerializer(nodes, marks);
   private schema: Schema;
+  private options: TransformerOptions;
 
-  constructor(schema: Schema = bitbucketSchema) {
+  constructor(
+    schema: Schema = bitbucketSchema,
+    options: TransformerOptions = {},
+  ) {
     this.schema = schema;
+    this.options = options;
   }
 
   encode(node: PMNode): string {
@@ -30,6 +31,6 @@ export default class BitbucketTransformer implements Transformer<string> {
   }
 
   buildDOMTree(html: string): HTMLElement {
-    return transformHtml(html);
+    return transformHtml(html, this.options);
   }
 }