+++ packages/fabric/editor-core/src/ui/ToolbarAdvancedTextFormatting/styles.ts
@@ -1,3 +1,4 @@
+import { akColorN30 } from '@atlaskit/util-shared-styles';
 import styled from 'styled-components';
 
 // tslint:disable-next-line:variable-name
@@ -6,6 +7,19 @@ export const TriggerWrapper = styled.div`
 `;
 
 // tslint:disable-next-line:variable-name
-export const ExpandIconWrapper = styled.div`
-  margin-left: -8px;
+export const Separator = styled.span`
+  background: ${akColorN30};
+  width: 1px;
+  height: 24px;
+  display: inline-block;
+  margin: 0 8px;
+`;
+
+// tslint:disable-next-line:variable-name
+export const Wrapper = styled.span`
+  display: flex;
+  align-items: center;
+  > div > div {
+    display: flex;
+  }
 `;