+++ packages/fabric/renderer/__tests__/browser/validator.ts
@@ -644,10 +644,10 @@ describe('Renderer - Validator', () => {
           attrs: {
             title: 'title',
             target: {
-              key: 'somу-key'
+              key: 'somу-key',
             },
-            parameters: {}
-          }
+            parameters: {},
+          },
         };
         expect(getValidMark(data)).to.deep.equal(data);
       });
@@ -661,7 +661,9 @@ describe('Renderer - Validator', () => {
       });
 
       it('should return null if attrs.target.key is missing', () => {
-        expect(getValidMark({ type: 'action', attrs: { target: {} }})).to.equal(null);
+        expect(
+          getValidMark({ type: 'action', attrs: { target: {} } }),
+        ).to.equal(null);
       });
     });
 
@@ -935,6 +937,41 @@ describe('Renderer - Validator', () => {
         expect(attrs.state).to.equal('DONE');
         expect(attrs.localId).to.not.equal(undefined);
         expect(content).to.deep.equal(itemContent);
+      });
+    });
+
+    describe('image', () => {
+      it('should pass through attrs as image', () => {
+        const imageAttrs = {
+          src: 'http://example.com/test.jpg',
+          alt: 'explanation',
+          title: 'image',
+        };
+        const { type, attrs } = getValidNode({
+          type: 'image',
+          attrs: imageAttrs,
+        });
+        expect(type).to.equal('image');
+        expect(attrs).to.deep.equal(imageAttrs);
+      });
+
+      it('should pass through attrs with only src as image', () => {
+        const imageAttrs = { src: 'http://example.com/test.jpg' };
+        const { type, attrs } = getValidNode({
+          type: 'image',
+          attrs: imageAttrs,
+        });
+        expect(type).to.equal('image');
+        expect(attrs).to.deep.equal(imageAttrs);
+      });
+
+      it('should reject image without src', () => {
+        const imageAttrs = { alt: 'explanation' };
+        const { type, attrs } = getValidNode({
+          type: 'image',
+          attrs: imageAttrs,
+        });
+        expect(type).to.equal('text');
       });
     });