+++ packages/elements/layer-manager/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/layer-manager
 
+## 2.5.0
+- [minor] support context via HOC from layer-manager [333a8de](333a8de)
+
+## 2.4.1
+
 ## 2.4.0 (2017-11-14)
 
 * add flow types