+++ packages/fabric/editor-common/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaskit/editor-common
 
+## 3.9.2
+
+## 3.9.1
+- [patch] Restore accessLevel attribute for mention node [a83619f](a83619f)
+- [patch] Restore accessLevel attribute for mention node [a83619f](a83619f)
+
+## 3.9.0
+
+## 3.8.3
+- [patch] Use correct dependencies  [7b178b1](7b178b1)
+- [patch] Use correct dependencies  [7b178b1](7b178b1)
+- [patch] Adding responsive behavior to the editor. [e0d9867](e0d9867)
+- [patch] Adding responsive behavior to the editor. [e0d9867](e0d9867)
+
 ## 3.8.2
 
 ## 3.8.1