+++ packages/fabric/renderer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaskit/renderer
 
+## 8.10.1
+
+## 8.10.0
+- [minor] Adding 'image' node for bitbucket consumption; this is unstable and should not be used [590ce41](590ce41)
+- [minor] Adding 'image' node for bitbucket consumption; this is unstable and should not be used [590ce41](590ce41)
+
+## 8.9.0
+
+## 8.8.1
+- [patch] Use correct dependencies  [7b178b1](7b178b1)
+- [patch] Use correct dependencies  [7b178b1](7b178b1)
+- [patch] Adding responsive behavior to the editor. [e0d9867](e0d9867)
+- [patch] Adding responsive behavior to the editor. [e0d9867](e0d9867)
+
 ## 8.8.0
 - [minor] Added big emoji rendering logic to renderer [f85c47a](f85c47a)
 - [minor] Added big emoji rendering logic to renderer [f85c47a](f85c47a)